seed:required was failing due to order of seeding
